# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMRS RSUD Dr. H. Chasan Boesoirie Ternate - API Documentation",
...
        "contact": {
            "name": "IT SIMRS RSUD Dr. H. Chasan Boesoirie",
            "url": "https://chasanboesoirie.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/service/medifirst2000",
            "description": "Server SIMRS (Proxy Anti-CORS - Aktif Otomatis)"
        },
        {
            "url": "https://chasanboesoirie.id/service/medifirst2000",
            "description": "Server Langsung (chasanboesoirie.id - Butuh Postman/No-CORS)"
        }
    ],
    "paths": {
...
        "/kiosk/get-tarif": {
            "get": {
                "tags": [
                    "Master Tarif & Harga Tindakan per Kelas"
                ],
                "summary": "Daftar Tarif & Harga Tindakan Lengkap per Kelas",
                "description": "Mengambil daftar seluruh tarif tindakan rumah sakit yang dikelompokkan berdasarkan Kelas Perawatan (Kelas 1, 2, 3, VIP, VVIP, Non Kelas), Ruangan/Poli, dan Jenis Pelayanan. Mendukung pencarian nama tindakan, filter produk, ruangan, kelas, dan jenis pelayanan.",
                "operationId": "11ff16416c71df17c85e96aec8eac9ff",
                "parameters": [
                    {
                        "name": "namaproduk",
                        "in": "query",
                        "description": "Cari berdasarkan nama tindakan / layanan (contoh: EKG, Konsul, USG)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "EKG"
                        }
                    },
                    {
                        "name": "kelasId",
                        "in": "query",
                        "description": "Filter ID Kelas Perawatan (1: Kelas 1, 2: Kelas 2, 3: Kelas 3, 4: VIP, 5: VVIP, 6: Non Kelas)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "ruanganId",
                        "in": "query",
                        "description": "Filter ID Ruangan Pelayanan",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "produkId",
                        "in": "query",
                        "description": "Filter ID Produk / Tindakan Spesifik",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "jenispelayananId",
                        "in": "query",
                        "description": "Filter ID Jenis Pelayanan",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar tarif tindakan berhasil dimuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID Tindakan (produk_m)",
                                                "type": "integer",
                                                "example": 2041
                                            },
                                            "namaproduk": {
                                                "type": "string",
                                                "example": "Elektrokardiogram (EKG)"
                                            },
                                            "hargalayanan": {
                                                "description": "Harga / Tarif Layanan",
                                                "type": "number",
                                                "format": "float",
                                                "example": 65000
                                            },
                                            "idkelas": {
                                                "description": "ID Kelas",
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "namakelas": {
                                                "description": "Nama Kelas Perawatan",
                                                "type": "string",
                                                "example": "Kelas III"
                                            },
                                            "jenispelayananid": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "jenispelayanan": {
                                                "type": "string",
                                                "example": "Rawat Jalan"
                                            },
                                            "ruid": {
                                                "type": "integer",
                                                "example": 45
                                            },
                                            "namaruangan": {
                                                "type": "string",
                                                "example": "Poli Jantung"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/humas/get-daftar-tarif-layanan": {
            "get": {
                "tags": [
                    "Master Tarif & Harga Tindakan per Kelas"
                ],
                "summary": "Informasi Publik Tarif Layanan Rumah Sakit",
                "description": "Menampilkan informasi tarif layanan dan tindakan untuk keperluan informasi publik dan humas rumah sakit.",
                "operationId": "70fa1387fc297aed74129294159a519a",
                "parameters": [
                    {
                        "name": "namaproduk",
                        "in": "query",
                        "description": "Nama tindakan yang dicari",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Darah Lengkap"
                        }
                    },
                    {
                        "name": "kelasId",
                        "in": "query",
                        "description": "ID Kelas Perawatan",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar tarif layanan humas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sysadmin/general/get-komponenharga": {
            "get": {
                "tags": [
                    "Master Tarif & Harga Tindakan per Kelas"
                ],
                "summary": "Rincian Komponen Harga & Tarif Tindakan",
                "description": "Mengambil pecahan rincian komponen pembentuk harga tindakan (Jasa Medis Dokter, Jasa Perawat/Paramedis, Jasa Sarana Rumah Sakit, Bahan Habis Pakai/BHP, Administrasi) berdasarkan ID Tindakan, ID Kelas, ID Ruangan, dan ID Jenis Pelayanan.",
                "operationId": "2a0d5817b703d4db11d1c502e5ab74a4",
                "parameters": [
                    {
                        "name": "idProduk",
                        "in": "query",
                        "description": "ID Tindakan / Produk (produk_m)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1042
                        }
                    },
                    {
                        "name": "idKelas",
                        "in": "query",
                        "description": "ID Kelas Perawatan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "idRuangan",
                        "in": "query",
                        "description": "ID Ruangan Pelayanan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "idJenisPelayanan",
                        "in": "query",
                        "description": "ID Jenis Pelayanan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rincian komponen harga tindakan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "objectkomponenhargafk": {
                                                "type": "integer",
                                                "example": 25
                                            },
                                            "komponenharga": {
                                                "type": "string",
                                                "example": "Jasa Medis Spesialis"
                                            },
                                            "hargasatuan": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 50000
                                            },
                                            "iscito": {
                                                "type": "boolean",
                                                "example": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Token otorisasi SIMRS Ternate. Default token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJoaXMuamtuIn0.FhnuCCWQx9SOMwBcLx3NFMPH45-KUwkblNKQY9Debm26PJ7ygt4-1z7oSMrefb0qfJwKtp02kS1O0lupFcMz1Q",
                "name": "X-AUTH-TOKEN",
                "in": "header"
            }
        }
    },
...
}